refactor: fold history into lib/last-location
`lib/history.ts` held one const and a one-line helper. Both are about the
app's current location, which `lib/last-location.ts` already covers, and
that module is a leaf both `main.tsx` and `lib/auth.ts` already import — so
the merge adds no edge and cannot close a cycle.
