+++ .agents/rules/architecture.md
@@ -28,7 +28,7 @@ Every app's entry imports `@tamery/shared/arktype-config` first (`exactOptionalP
 - `lib/database.ts` opens the OPFS wa-sqlite database in a **top-level await** — `lib/sync.ts`, every collection, and anything transitively importing it waits for WASM + OPFS before evaluating. Anything reachable from `main.tsx`/`routeTree.gen.ts` blocks first paint, for signed-out users too.
 - Route modules are all imported eagerly by the generated tree; the splitter only moves `component`/`loader`/`*Component`, so `beforeLoad` and its imports stay eager. Keep the data layer off that path: reach collections through the parent route's context, `await import()` them inside `beforeLoad`, and import the leaf module that owns the symbol — `entities/*` has no barrels to re-export `fetching` and drag kysely in (`code-style.md`).
 - `entities/connection` must stay acyclic: `queries/*` call `createQuery` at module scope, so a cycle back into `runtime/query.ts` surfaces as `Cannot access 'createQuery' before initialization` at whichever module the bundler enters the cycle from. `fetchingConfig` therefore lives in the leaf `utils/fetching-config.ts`; `runtime/dialects` imports it from there, never from `utils/fetching.ts` (which reaches `queries/*`). `import/no-cycle` enforces this repo-wide (`code-style.md`).
-- `main.tsx` exports nothing — it only creates the router and renders. Shared singletons live in leaves so no route or `lib/*` module imports the entry: query clients in `lib/query-client.ts`, the router history in `lib/history.ts`. Library code that needs the current URL or a redirect uses that history (`isAuthLocation()`, `history.push`), never the router instance; components use `useRouter()`.
+- `main.tsx` exports nothing — it only creates the router and renders. Shared singletons live in leaves so no route or `lib/*` module imports the entry: query clients in `lib/query-client.ts`, the router history in `lib/last-location.ts`. Library code that needs the current URL or a redirect uses that history (`isAuthLocation()`, `history.push`), never the router instance; components use `useRouter()`.
 - The window paints app chrome before any of that: `src/shell.tsx` is prerendered into `index.html` by `@tamery/vite-prerender`, wired in `vite.config.ts` with component/marker pairs (dev and build alike). Keep it hook-free and Node-safe; design rules in the `tamery-ui` skill.
 - `src/lib/warmup.ts` is the entry's first import and the only module that deliberately kicks off heavy chunks — `import()` of the database right away, monaco 1s after `load`. Monaco and `lib/database` must never be static-imported from the entry; posthog-js stays behind the lazy `lib/posthog.ts` facade.
 - Verify by walking the dev module graph from `/src/main.tsx`, not by reading imports. The `@hugeicons/core-free-icons` barrel via `packages/ui` `sonner.tsx` is eager in dev only (prod treeshakes it).
