+++ apps/app/src/lib/last-location.ts
@@ -1,7 +1,16 @@
+import { createBrowserHistory, createHashHistory } from '@tanstack/react-router'
 import { type } from 'arktype'
 import { createWebStorageValue } from 'seitu/web'
 
 import { LAST_LOCATION_KEY } from './constants'
+
+export const history =
+  import.meta.env.VITE_TEST || !window.electron
+    ? createBrowserHistory()
+    : createHashHistory()
+
+export const isAuthLocation = () =>
+  history.location.pathname.startsWith('/auth')
 
 export const lastLocationStorageValue = createWebStorageValue({
   defaultValue: null,
